+++ projects/bridge-lz/index.js
@@ -1,0 +1,17 @@
+const { sumTokensExport, nullAddress } = require("../helper/unwrapLPs");
+
+const ADMIN_ADDRESSES = {
+  base: "0x233dc79F924c35AcB4524BaC4A883c8CE11A75B2",
+  optimism: "0x233dc79F924c35AcB4524BaC4A883c8CE11A75B2",
+}
+
+module.exports = {
+  methodology:
+    "Adds up the total value locked as collateral on the Bridge platform"
+}
+
+Object.keys(ADMIN_ADDRESSES).forEach(chain => {
+  module.exports[chain] = {
+    tvl: sumTokensExport({ owner: ADMIN_ADDRESSES[chain], tokens: [nullAddress]})
+  }
+})
